requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==22.1.0
boto3==1.26.35
botocore==1.29.68
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
Django==3.2.6
django-activity-stream==1.4.1
django-cors-headers==3.11.0
django-environ==0.8.1
django-extensions==3.1.5
django-filter==22.1
django-location-field==2.1.0
django-mapbox-location-field==2.0.0
django-rest-knox==4.2.0
django-storages==1.13.2
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
drf-extra-fields==3.4.0
drf-yasg2==1.19.4
Faker==16.8.1
frozenlist==1.3.1
idna==3.3
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.1
jmespath==1.0.1
MarkupSafe==2.1.1
maxminddb==2.2.0
middleware==1.2.3
multidict==6.0.2
oauthlib==3.2.0
packaging==21.3
Pillow==9.1.0
psycopg2==2.9.3
pycparser==2.21
pydot==1.4.2
PyJWT==2.3.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
yarl==1.8.1
